+++ Firestore/Source/API/FSTUserDataWriter.mm
@@ -37,7 +37,6 @@
 @class FIRTimestamp;
 
 namespace api = firebase::firestore::api;
-namespace util = firebase::firestore::util;
 namespace model = firebase::firestore::model;
 namespace nanopb = firebase::firestore::nanopb;
 
@@ -49,6 +48,7 @@ using firebase::firestore::google_firestore_v1_ArrayValue;
 using firebase::firestore::google_firestore_v1_MapValue;
 using firebase::firestore::google_firestore_v1_Value;
 using firebase::firestore::google_protobuf_Timestamp;
+using firebase::firestore::util::MakeNSString;
 using model::DatabaseId;
 using model::DocumentKey;
 using model::GetLocalWriteTime;
@@ -99,7 +99,7 @@ NS_ASSUME_NONNULL_BEGIN
                  ? @(value.integer_value)
                  : @(value.double_value);
     case TypeOrder::kString:
-      return util::MakeNSString(MakeStringView(value.string_value));
+      return MakeNSString(MakeStringView(value.string_value));
     case TypeOrder::kBlob:
       return MakeNSData(value.bytes_value);
     case TypeOrder::kGeoPoint:
@@ -115,7 +115,7 @@ NS_ASSUME_NONNULL_BEGIN
   for (pb_size_t i = 0; i < mapValue.fields_count; ++i) {
     absl::string_view key = MakeStringView(mapValue.fields[i].key);
     const google_firestore_v1_Value &value = mapValue.fields[i].value;
-    result[util::MakeNSString(key)] = [self convertedValue:value];
+    result[MakeNSString(key)] = [self convertedValue:value];
   }
   return result;
 }