+++ FirebaseDynamicLinks/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v8.8.0
+- [fixed] Firebase dynamic links with custom domain will only work if the custom domain has a trailing '/'. (#7087)
+- [fixed] Fix device-only build warning for unused `processIsTranslated` function. (#8694)
+
 # v8.7.0
 - [added] Refactoring and adding helper class. (#8432)